requirements.txt
# HuggingFace Space — CPU/ZeroGPU uyumlu
gradio>=4.44.0
transformers>=4.45.0
accelerate>=0.34.0
opencv-python-headless>=4.10.0
httpx>=0.27.0
Pillow>=10.0.0
